requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
decouple==0.0.7
dj-database-url==1.0.0
dj-stripe==2.7.2
Django==4.1.3
django-bootstrap-form==3.4
django-ckeditor==6.5.1
django-filter==22.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-on-heroku==1.1.2
django-tinymce==3.5.0
fontawesomefree==6.0.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
jsonfield==3.1.0
numpy==1.24.0
pandas==1.5.2
Pillow==9.3.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stripe==5.0.0
urllib3==1.26.12
whitenoise==6.2.0
